Move safe_float to own utils module
