+++ training/object-detection/src/object_detection/utils/utils.py
@@ -1,0 +1,5 @@
+def safe_float(x: str, default: float = 0.0) -> float:
+    try:
+        return float(x)
+    except Exception:
+        return default
